+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Build releases using Go v1.26
 - Bump dependencies
 - Allow sprite frames to have different sizes
+- Fix missing color in index-alpha sprite extract
 
 # v1.5.0
 - Add "map neat" entity preprocessor command